requirement.txt
Flask==2.1.0
Flask-SQLAlchemy==3.0.0
boto3==1.18.0
numpy==1.21.0